Move styles for SortingTableHEad to its own style file
