+++ context/app/static/js/components/Search/SortingTableHead/style.js
@@ -1,5 +1,9 @@
 import styled from 'styled-components';
 import TableHead from '@material-ui/core/TableHead';
+import ArrowUpward from '@material-ui/icons/ArrowUpwardRounded';
+import ArrowDownward from '@material-ui/icons/ArrowDownwardRounded';
+
+import { HeaderCell } from 'js/shared-styles/Table';
 
 const StyledTableHead = styled(TableHead)`
   ${({ searchView }) =>
@@ -9,4 +13,27 @@ const StyledTableHead = styled(TableHead)`
   `}
 `;
 
-export { StyledTableHead };
+const ArrowUpOn = styled(ArrowUpward)`
+  vertical-align: middle;
+`;
+
+const ArrowDownOn = styled(ArrowDownward)`
+  vertical-align: middle;
+`;
+
+const ArrowUpOff = styled(ArrowUpward)`
+  vertical-align: middle;
+  opacity: 12%;
+`;
+
+const ArrowDownOff = styled(ArrowDownward)`
+  vertical-align: middle;
+  opacity: 12%;
+`;
+
+const StyledHeaderCell = styled(HeaderCell)`
+  cursor: pointer;
+  white-space: nowrap;
+`;
+
+export { StyledTableHead, ArrowUpOn, ArrowDownOn, ArrowUpOff, ArrowDownOff, StyledHeaderCell };
